+++ scanner/src/main/java/no/nordicsemi/android/common/scanner/viewmodel/ScannerViewModel.kt
@@ -106,6 +106,7 @@ internal class ScannerViewModel @Inject constructor(
                         showNonEmptyName = true,
                         showBonded = true,
                         showSortByOption = true,
+                        showPeripheralDropdown = true,
                     )
                 )
             )
